# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Movies API",
    "version": "1.0.0",
    "description": "API para gestión de películas con operaciones CRUD",
    "contact": {
      "name": "Erick Verde"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de desarrollo"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Ingresa el token en el formato: 1234"
      }
    },
    "schemas": {
      "Movie": {
        "type": "object",
        "required": [
          "id",
          "title",
          "year",
          "genre",
          "director",
          "actors",
          "plot",
          "imdb_rating",
          "runtime_minutes"
        ],
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^tt\\d+$",
            "description": "ID único de IMDb (formato: ttXXXXXXX)",
            "example": "tt0111161"
          },
          "title": {
            "type": "string",
            "description": "Título de la película",
            "example": "The Shawshank Redemption"
          },
          "year": {
            "type": "integer",
            "minimum": 1900,
            "maximum": 2024,
            "description": "Año de lanzamiento",
            "example": 1994
          },
          "genre": {
            "type": "string",
            "description": "Géneros separados por comas",
            "example": "Drama, Crime"
          },
          "director": {
            "type": "string",
            "description": "Director de la película",
            "example": "Frank Darabont"
          },
          "actors": {
            "type": "string",
            "description": "Actores principales separados por comas",
            "example": "Tim Robbins, Morgan Freeman"
          },
          "plot": {
            "type": "string",
            "description": "Sinopsis de la película",
            "example": "Two imprisoned men bond over a number of years, finding solace and eventual redemption through acts of common decency."
          },
          "imdb_rating": {
            "type": "number",
            "minimum": 0,
            "maximum": 10,
            "description": "Calificación de IMDb",
            "example": 9.3
          },
          "runtime_minutes": {
            "type": "integer",
            "minimum": 1,
            "description": "Duración en minutos",
            "example": 142
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/movies": {
      "get": {
        "summary": "Obtener todas las películas",
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Filtrar por limite de resultados , por defecto 10"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Filtrar por página de resultados, por defecto 1"
          },
          {
            "in": "query",
            "name": "orden",
            "schema": {
              "type": "string"
            },
            "description": "Filtrar por orden de resultados, valores year o rating"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de películas obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "estado": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Movie"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Agregar una nueva película",
        "tags": [
          "Movies - CRUD"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Película creada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "estado": {
                      "type": "string",
                      "example": "success"
                    },
                    "mensaje": {
                      "type": "string",
                      "example": "Película agregada exitosamente"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Movie"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Token no proporcionado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Token inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Película duplicada (ID o título ya existe)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/movies/year": {
      "get": {
        "summary": "Obtener películas por rango de años (query parameter)",
        "tags": [
          "Movies - Filter"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "schema": {
              "type": "integer"
            },
            "description": "Año desde (mayor que)"
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "type": "integer"
            },
            "description": "Año hasta (menor que)"
          }
        ],
        "responses": {
          "200": {
            "description": "Películas filtradas por año",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "estado": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Movie"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/movies/year/{year}": {
      "get": {
        "summary": "Obtener películas por año específico",
        "tags": [
          "Movies - Filter"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "year",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Año de la película",
            "example": 2020
          }
        ],
        "responses": {
          "200": {
            "description": "Películas del año especificado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "estado": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Movie"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/movies/director/{name}": {
      "get": {
        "summary": "Buscar películas por director específico",
        "tags": [
          "Movies - Search"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Nombre del director",
            "example": "Christopher Nolan"
          }
        ],
        "responses": {
          "200": {
            "description": "Películas encontradas del director",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "estado": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Movie"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/movies/actor/{name}": {
      "get": {
        "summary": "Buscar películas por actor específico",
        "tags": [
          "Movies - Search"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Nombre del actor",
            "example": "Leonardo DiCaprio"
          }
        ],
        "responses": {
          "200": {
            "description": "Películas encontradas con el actor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "estado": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Movie"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/movies/rating": {
      "get": {
        "summary": "Filtrar películas por rango de calificación",
        "tags": [
          "Movies - Filter"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "schema": {
              "type": "number"
            },
            "description": "Calificación mínima"
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "type": "number"
            },
            "description": "Calificación máxima"
          }
        ],
        "responses": {
          "200": {
            "description": "Películas filtradas por calificación",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "estado": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Movie"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}